added 0_common/50_set_OS-version.sh: sets $OS_VERSION to 5 or 6 depending on return of t_DistCheck changed p_git/0-install_git.sh: Test is skipped in C5 changed p_git/10-test_git.sh: made test more verbose (added t_Log), skipping on C5 changed p_git/11-test_git_clone.sh: made test more verbose (added t_log), test now ends in the original dir (previously it ended in /tmp, skipping on C5 changed p_git/12-test_git_log.sh: made test more verbose (added t_log), test now ends in the original dir (pr eviously it ended in /tmp, skipping on C5

+++ tests/p_git/12-test_git_log.sh
@@ -1,18 +1,31 @@
 #!/bin/bash
 
 # Author: Pratima Singh <prati.86@gmail.com>, Nilesh Bairagi <nileshbairagi@gmail.com>, Madhurranjan Mohaan <madhurranjan.mohaan@gmail.com>
+# 	  Christoph Galuschka <christoph.galuschka@chello.at>
 
-# Tsting git log and git commit msgs
-rm -rf /tmp/temprepo/ /tmp/cloned_repo/ 
-temp_repo="/tmp/temprepo"
-mkdir -p $temp_repo
-cd $temp_repo
-git init . --bare
-cd /tmp
-git clone $temp_repo cloned_repo
-cd cloned_repo
-echo "hello world" > hello
-git add hello
-git commit -m "Temp commit" 2>&1
-git log --grep="Temp commit" 2>&1
-t_CheckExitStatus $?
+t_Log "Running $0 - testing git logging and commit messages"
+
+if [ $OS_VERSION -eq '6' ]
+  then
+  # Testing git log and git commit msgs
+  workpath=$(pwd)
+  rm -rf /tmp/temprepo/ /tmp/cloned_repo/ 
+  temp_repo="/tmp/temprepo"
+  mkdir -p $temp_repo
+  cd $temp_repo
+  git init . --bare
+  cd /tmp
+  git clone $temp_repo cloned_repo
+  cd cloned_repo
+  echo "hello world" > hello
+  git add hello
+  git commit -m "Temp commit" 2>&1
+  git log --grep="Temp commit" 2>&1
+  ret_val=$?
+  cd $workpath
+else
+  t_Log "This test is skipped in CentOS5."
+  ret_val=0
+fi
+
+t_CheckExitStatus $ret_val
